Fix: Long to Int conversion in DrawingCanvas.kt
- Add .toInt() to all hex color values in colorPalette
- Hex values > Int.MAX_VALUE were inferred as Long
- This caused compile error: 'Type mismatch: inferred type is Long but Int was expected'
- Also fix default color in Stroke data class

+++ app/src/main/java/com/arwallcanvas/DrawingCanvas.kt
@@ -31,14 +31,14 @@ class DrawingCanvas(private val session: Session) {
 
     // Cores para alternar
     private val colorPalette = listOf(
-        0xFFFF4444, // Vermelho
-        0xFF44FF44, // Verde
-        0xFF4444FF, // Azul
-        0xFFFFFF44, // Amarelo
-        0xFFFF44FF, // Rosa
-        0xFF44FFFF, // Ciano
-        0xFFFFFFFF, // Branco
-        0xFFFF8800  // Laranja
+        0xFFFF4444.toInt(), // Vermelho
+        0xFF44FF44.toInt(), // Verde
+        0xFF4444FF.toInt(), // Azul
+        0xFFFFFF44.toInt(), // Amarelo
+        0xFFFF44FF.toInt(), // Rosa
+        0xFF44FFFF.toInt(), // Ciano
+        0xFFFFFFFF.toInt(), // Branco
+        0xFFFF8800.toInt()  // Laranja
     )
     private var colorIndex = 0
